Fixing an issue with string-based transformations after a change in Uglify
In Uglify 2.4.17, there was a change in the replacements done in the
source text before it is tokenized, resulting in differences between the
positions reported by Uglify and the ones understood by at-noder-converter.

Instead of reproducing the same replacements, at-noder-converter now
directly calls the tokenizer to be sure it uses the same sourceText,
so that changes using positions in the string, as performed by
insertNodeInString, are accurate.

See the following commit in UglifyJS:
https://github.com/mishoo/UglifyJS2/commit/13219cebcb11f95fdaa77af653e96ac42e7e6dcc